Fix critical security vulnerabilities and budget blind spots
