+++ src/coreason_manifest/adapters/observability/integrity.py
@@ -24,7 +24,7 @@ def to_canonical_timestamp(dt: datetime) -> str:
     dt_utc = dt.astimezone(UTC)
 
     # Format as YYYY-MM-DDTHH:MM:SSZ (no microseconds)
-    return dt_utc.strftime("%Y-%m-%dT%H:%M:%SZ")
+    return dt_utc.strftime("%Y-%m-%dT%H:%M:%S.%fZ")
 
 
 class MerkleNode(TypedDict):
