+++ src/coreason_manifest/core/compliance.py
@@ -1,7 +1,10 @@
+import ast
 from enum import StrEnum
 from typing import Any, Literal
 
 from pydantic import BaseModel, ConfigDict, Field
+
+from coreason_manifest.core.exceptions import ManifestError, ManifestErrorCode
 
 
 class ErrorCatalog(StrEnum):
@@ -81,3 +84,61 @@ class ComplianceReport(BaseModel):
     node_id: str | None = None
     remediation: RemediationAction | None = None
     details: dict[str, Any] = Field(default_factory=dict)
+
+
+class SecurityVisitor(ast.NodeVisitor):
+    """
+    Centralized strict AST whitelisting to prevent code injection via eval/exec interfaces.
+    Enforces a strict subset of safe Python expression nodes.
+    """
+
+    def generic_visit(self, node: ast.AST) -> None:
+        # Whitelist of allowed AST nodes (Tuple for fast isinstance checks)
+        allowed = (
+            ast.Expression,
+            ast.BoolOp,
+            ast.BinOp,
+            ast.UnaryOp,
+            ast.Compare,
+            ast.Constant,
+            ast.Name,
+            ast.Load,
+            ast.And,
+            ast.Or,
+            ast.Eq,
+            ast.NotEq,
+            ast.Lt,
+            ast.LtE,
+            ast.Gt,
+            ast.GtE,
+            ast.Is,
+            ast.IsNot,
+            ast.In,
+            ast.NotIn,
+            ast.Not,
+            ast.Add,
+            ast.Attribute,
+            ast.Sub,
+            ast.Mult,
+            ast.Div,
+            ast.FloorDiv,
+            ast.Mod,
+            ast.Pow,
+            ast.USub,
+            ast.UAdd,
+        )
+        if not isinstance(node, allowed):
+            raise ManifestError.critical_halt(
+                code=ManifestErrorCode.CRSN_SEC_KILL_SWITCH_VIOLATION,
+                message=f"Security Violation: forbidden AST node {type(node).__name__} in evaluated condition",
+            )
+        super().generic_visit(node)
+
+    def visit_Name(self, node: ast.Name) -> None:
+        # Ensure Name usage is strictly Load context (prevents variable reassignment)
+        if not isinstance(node.ctx, ast.Load):
+            raise ManifestError.critical_halt(
+                code=ManifestErrorCode.CRSN_SEC_KILL_SWITCH_VIOLATION,
+                message=f"Security Violation: Name ctx {type(node.ctx).__name__} forbidden in evaluated condition",
+            )
+        super().generic_visit(node)
